+++ core-ui/src/components/Predefined/Create/Jobs/CronJobs.create.js
@@ -2,22 +2,12 @@ import React, { useEffect, useState } from 'react';
 import * as jp from 'jsonpath';
 import { useTranslation } from 'react-i18next';
 
-import * as Inputs from 'shared/ResourceForm/components/Inputs';
 import { ResourceForm } from 'shared/ResourceForm/ResourceForm';
+
 import { createCronJobTemplate, createCronJobPresets } from './templates';
-import { SpecSection } from './SpecSection';
+import { CronJobSpecSection } from './SpecSection';
 import { isCronExpressionValid, ScheduleSection } from './ScheduleSection';
 import { ContainerSection, ContainersSection } from './ContainersSection';
-
-const concurrencyPolicyOptions = ['Allow', 'Forbid', 'Replace'].map(p => ({
-  key: p,
-  text: p,
-}));
-
-const restartPolicyOptions = ['Never', 'OnFailure'].map(p => ({
-  key: p,
-  text: p,
-}));
 
 function isCronJobValid(cronJob) {
   const containers =
@@ -81,7 +71,7 @@ export function CronJobsCreate({
         title={t('common.headers.annotations')}
       />
 
-      <SpecSection
+      <CronJobSpecSection
         advanced
         resource={cronJob.spec}
         setResource={spec => {
@@ -89,24 +79,6 @@ export function CronJobsCreate({
           setCronJob({ ...cronJob });
         }}
       />
-
-      <ResourceForm.FormField
-        advanced
-        propertyPath="$.spec.concurrencyPolicy"
-        label={t('cron-jobs.create-modal.labels.concurrency-policy')}
-        input={Inputs.Dropdown}
-        defaultKey="Allow"
-        options={concurrencyPolicyOptions}
-      />
-
-      <ResourceForm.FormField
-        advanced
-        propertyPath="$.spec.jobTemplate.spec.template.spec.restartPolicy"
-        label={t('cron-jobs.create-modal.labels.restart-policy')}
-        input={Inputs.Dropdown}
-        options={restartPolicyOptions}
-      />
-
       <ScheduleSection propertyPath="$.spec.schedule" />
 
       <ContainerSection
